requirements.txt
# requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
sqlalchemy==1.4.23
psycopg2-binary==2.9.1
pydantic==1.8.2
python-multipart==0.0.5
stripe==2.60.0
alembic==1.7.1
python-jose==3.3.0
passlib==1.7.4
python-dotenv==0.19.0
httpx==0.18.2
python-docx==0.8.11
pypdf2==1.26.0
